requirements.txt
openai==0.27.0  # OpenAI Python SDK
flask==2.2.2    # Flask用来搭建API接口
requests==2.28.1 # 如果需要请求其他音频生成服务的API
python-dotenv==0.19.2 # 用于读取环境变量（如API密钥）
